**refactor(isobmff): normalize box/length math to `int` and align readers/tests**
- convert ISOBMFF size/remain/length handling from mixed `int64`/`uint*` to `int` across box parsing, metadata traversal, CRX, PRVW, and XPacket paths
- update `XPacketHeader.Length` and related call sites/logging/tests to use `int`
- add explicit bounded conversions (`uint64ToInt`) and keep overflow checks centralized
- refresh ISOBMFF tests for new max-int behavior and remove redundant loop-var copies
- add targeted `nolint:gosec` annotations where casts are spec-bounded

**chore(lint): fix root benchmark/reader lint findings**

- add checked pooled-reader helper in `imagemeta.go` to satisfy `errcheck` type-assertion rules
- replace direct error equality checks with `errors.Is` in benchmarks (`errorlint`)
- remove benchmark shadowed `err` variables (`govet`) and clean pooled-reader error paths

+++ meta/isobmff/bmff.go
@@ -20,7 +20,7 @@ type PreviewImageReader func(r io.Reader, h meta.PreviewHeader) error
 // XPacketHeader describes discovered XPacket payload metadata.
 type XPacketHeader struct {
 	Offset       uint64
-	Length       uint32
+	Length       int
 	HasXPacketPI bool
 	HasXMPMeta   bool
 }
@@ -80,12 +80,12 @@ func parseFileTypeBox(b *box) (ftyp fileTypeBox, err error) {
 		return ftyp, fmt.Errorf("parseFileTypeBox: %w", ErrBufLength)
 	}
 
-	peekLen := int64(ftypHeaderSize + (fourCCSize * maxBrandCount))
+	peekLen := ftypHeaderSize + (fourCCSize * maxBrandCount)
 	if b.remain < peekLen {
 		peekLen = b.remain
 	}
 
-	buf, err := b.Peek(int(peekLen))
+	buf, err := b.Peek(peekLen)
 	if err != nil {
 		return ftyp, err
 	}
